Depurando la gestion de fotos de perfil, daba algunos errores

+++ src/routes/fotos.js
@@ -13,8 +13,8 @@ const storage = multer.diskStorage({
     destination: (req, file, cb) => {
         const { nif } = req.body;
         const { user } = req.body;
-          //Si usuario es undefined es que se subió una baliza, y configuro el storage para balizas
-          if (typeof user === 'undefined') {
+        //Si usuario es undefined es que se subió una baliza, y configuro el storage para balizas
+        if (typeof user === 'undefined') {
             const dir = path.join(__dirname, '../public/img/imagenes/', nif);
             fs.access(dir, error => {
                 if (error) {
@@ -25,12 +25,12 @@ const storage = multer.diskStorage({
                 }
                 return cb(null, dir);
             });
-          } else{//si no, entonces es una foto de perfil y va a otra carpeta
+        } else {//si no, entonces es una foto de perfil y va a otra carpeta
             const dir = path.join(__dirname, '../public/img/profiles/');
             //console.log("dir" + dir);
             return cb(null, dir);
         }
-       
+
     },
     filename: (req, file, cb) => {
         cb(null, new Date().getTime() + path.extname(file.originalname).toLowerCase());
@@ -56,8 +56,8 @@ const uploadFoto = multer({
 router.post("/aton/upload/:nif", uploadFoto, funciones.isAuthenticated, async (req, res) => {
     console.log("Subiendo foto baliza");
     const { nif } = req.params;
-    console.log(req.params);
-    console.log(req.body);
+    req.flash("success", "Nueva fotografia insertada correctamente");
+    funciones.insertarLog(req.user.usuario, "INSERT fotografia", nif);
     res.redirect("/aton/plantilla/" + nif);
 });
 router.get("/aton/fotos/:nif", async (req, res) => {
@@ -65,45 +65,50 @@ router.get("/aton/fotos/:nif", async (req, res) => {
     var fotos = funciones.listadoFotos(nif);
     res.render("aton/fotos", { fotos, nif });
 });
-router.get("/aton/fotos/:nif/:src/delete",  funciones.isAuthenticated, funciones.isAdmin, async (req, res) => {
+router.get("/aton/fotos/:nif/:src/delete", funciones.isAuthenticated, funciones.isAdmin, async (req, res) => {
     const nif = req.params.nif;
     const src = req.params.src;
     await unlink(path.resolve('src/public/img/imagenes/' + nif + "/" + src));
+    req.flash("success", "Fotografia borrada correctamente");
+    funciones.insertarLog(req.user.usuario, "DELETE fotografia", nif);
     res.redirect("/aton/fotos/" + nif);
 });
 
 //GESTION  foto perfil
-router.post('/upload/:id', funciones.isAuthenticated,uploadFoto, async (req, res) => {
+router.post('/upload/:id', funciones.isAuthenticated, uploadFoto, async (req, res) => {
     const { id } = req.params;
     //console.log(req.file);
     var usuario = await db.query("select * from usuarios where id = ?", id);
     usuario = usuario[0];
 
     //borramos la foto anterior del perfil
-    const filePath = path.resolve('src/public/img/profiles/' + usuario.pictureURL);
-    access(filePath, constants.F_OK, async (err) => {
-        if (err) {
-            req.flash("warning", "No tiene foto de perfil!");
-            console.log("No tiene foto de perfil");
-        } else {
-            console.log('File exists. Deleting now ...');
-            await unlink(filePath);
-        }
-    });
+    if (usuario.pictureURL != "") {
+        const filePath = path.resolve('src/public/img/profiles/' + usuario.pictureURL);
+        access(filePath, constants.F_OK, async (err) => {
+            if (err) {
+                req.flash("warning", "No tiene foto de perfil!");
+                console.log("No tiene foto de perfil");
+            } else {
+                console.log('File exists. Deleting now ...');
+                await unlink(filePath);
+            }
+        });
+    }
 
     //Ponemos la nueva
     usuario.pictureURL = req.file.filename;
     await db.query("UPDATE usuarios set  ? WHERE id=?", [usuario, id]);
+    funciones.insertarLog(req.user.usuario, "UPDATE fotografia perfil", "");
     req.flash("success", "Foto de perfil actualizada con exito");
     res.redirect("/profile");
 });
-router.get("/profile/borrarfoto/:id/:url", funciones.isAuthenticated, async(req, res) => {
+router.get("/profile/borrarfoto/:id/:url", funciones.isAuthenticated, async (req, res) => {
     //console.log(req.params);
     const { url } = req.params;
     const { id } = req.params;
     await db.query("UPDATE usuarios set pictureURL = NULL WHERE id=?", [id]);
     const filePath = path.resolve('src/public/img/profiles/' + url);
-    access(filePath, constants.F_OK, async(err) => {
+    access(filePath, constants.F_OK, async (err) => {
         if (err) {
             console.log("No tiene foto de perfil");
         } else {
@@ -111,6 +116,7 @@ router.get("/profile/borrarfoto/:id/:url", funciones.isAuthenticated, async(req,
             await unlink(filePath);
         }
     });
+    funciones.insertarLog(req.user.usuario, "DELETE fotografia perfil", "");
     req.flash("success", "Imagen borrada correctamente");
     res.redirect('/profile');
 });
